Refactor: 유저 레벨 업데이트 메서드 파라미터를 Double 타입으로 수정

+++ src/main/java/com/est/jobshin/domain/interview/service/InterviewService.java
@@ -2,7 +2,6 @@ package com.est.jobshin.domain.interview.service;
 
 import com.est.jobshin.domain.interview.domain.Interview;
 import com.est.jobshin.domain.interview.dto.InterviewDto;
-import com.est.jobshin.domain.interview.dto.InterviewLevel;
 import com.est.jobshin.domain.interview.repository.InterviewRepository;
 import com.est.jobshin.domain.interviewDetail.domain.InterviewDetail;
 import com.est.jobshin.domain.interviewDetail.dto.InterviewQuestion;
@@ -172,14 +171,14 @@ public class InterviewService {
                 .orElseThrow(() -> new NoSuchElementException("User not found"));
     }
 
-    public void updateUserLevel(Long score) {
+    public void updateUserLevel(Double score) {
         User user = getCurrentUser();
         Level newLevel = updateLevel(score);
         user.updateUserLevel(newLevel);
         userRepository.save(user);
     }
 
-    private Level updateLevel(Long score) {
+    private Level updateLevel(Double score) {
         if (score < 40) {
             return Level.LV1;
         } else if (score > 60) {
